filling loratap with bw, sf and freq

+++ include/lora/loratap.h
@@ -34,7 +34,7 @@ typedef enum sf { SF7=7, SF8, SF9, SF10, SF11, SF12 } sf_t;
 
 typedef struct __attribute__((__packed__)) loratap_channel {
 	uint32_t			frequency;	/* LoRa frequency (Hz) */
-	uint8_t				bandwidth;	/* Channel bandwidth (KHz) in 125 KHz steps */
+	uint32_t			bandwidth;	/* Channel bandwidth (KHz) in 125 KHz steps */
 	uint8_t				sf;		/* LoRa SF (sf_t) [7, 8, 9, 10, 11, 12] */
 } loratap_channel_t;
 
@@ -42,7 +42,7 @@ typedef struct __attribute__((__packed__)) loratap_rssi {
 	uint8_t				packet_rssi;	/* LoRa packet RSSI, if snr >= 0 then dBm value is -139 + packet_rssi, otherwise dBm value is -139 + packet_rssi * .25 */
 	uint8_t				max_rssi;	/* LoRa receiver max RSSI (dBm value is -139 + rssi) */
 	uint8_t				current_rssi;	/* LoRa receiver current RSSI (dBm value is -139 + rssi) */
-	uint8_t				snr;		/* LoRa SNR (dB value is (snr[two's complement])/4) */
+	uint32_t				snr;		/* LoRa SNR (dB value is (snr[two's complement])/4) */
 } loratap_rssi_t;
 
 typedef struct __attribute__((__packed__)) loratap_header {
